+++ src/cljs/wombats_web_client/components/arena_table.cljs
@@ -15,29 +15,34 @@
 
 
 (defn open-create-game-modal [arena-id]
-  (fn []
-    (re-frame/dispatch [:set-modal {:fn #(create-game-modal arena-id)
-                                    :show-overlay true}])))
+  (re-frame/dispatch [:set-modal {:fn #(create-game-modal arena-id)
+                                  :show-overlay true}]))
 
 (defn create-game-button [arena-id]
   [:input.create-game {:type "button"
                        :value "CREATE GAME"
-                       :on-click (open-create-game-modal arena-id)}])
+                       :on-click #(open-create-game-modal arena-id)}])
 
 (defn get-items-fn [row-data]
-  (let [name (:arena/name row-data)
-        width (:arena/width row-data)
-        height (:arena/height row-data)
-        food (:arena/food row-data)
-        poison (:arena/poison row-data)
-        steel (:arena/steel-walls row-data)
-        wood (:arena/wood-walls row-data)
-        zakano (:arena/zakano row-data)
-        dimension (str width "x" height)
-        id (:arena/id row-data)
-        create-game (create-game-button id)]
-
-    [name dimension food poison steel wood zakano create-game]))
+  (let [{:keys [:arena/width
+                :arena/height
+                :arena/food
+                :arena/poison
+                :arena/steel-walls
+                :arena/wood-walls
+                :arena/zakano
+                :arena/id
+                :arena/name]} row-data
+                dimension (str width "x" height)
+                create-game (create-game-button id)]
+    [name
+     dimension
+     food
+     poison
+     steel-walls
+     wood-walls
+     zakano
+     create-game]))
 
 (defn arena-table [arenas]
   [:div.arena-display