+++ src/cljs/wombats_web_client/utils/errors.cljs
@@ -10,6 +10,8 @@
   "This game has already started. Please try joining another game.")
 (defonce not-an-integer
   "Not an integer.")
+(defonce max-eight "Max is 8.")
+(defonce incorrect-format-colon "Format is mm:ss.")
 
 (def login-error "login-error")